(PC-13518)[PRO] refactor: move constants into core
