+++ pro/src/repository/pcapi/__specs__/pcapi.spec.js
@@ -1,4 +1,4 @@
-import { DEFAULT_PRE_FILTERS } from 'components/pages/Bookings/PreFilters/_constants'
+import { DEFAULT_PRE_FILTERS } from 'core/Bookings/constants'
 import { ALL_OFFERERS, DEFAULT_SEARCH_FILTERS } from 'core/Offers/constants'
 import {
   getBooking,
