+++ pro/src/repository/pcapi/pcapi.js
@@ -1,5 +1,5 @@
-import { DEFAULT_PRE_FILTERS } from 'components/pages/Bookings/PreFilters/_constants'
 import { DEFAULT_INVOICES_FILTERS } from 'components/pages/Reimbursements/_constants'
+import { DEFAULT_PRE_FILTERS } from 'core/Bookings'
 import { ALL_OFFERERS, DEFAULT_SEARCH_FILTERS } from 'core/Offers/constants'
 import { client } from 'repository/pcapi/pcapiClient'
 import {
