+++ pro/src/routes/Bookings/Bookings.tsx
@@ -6,7 +6,7 @@ import useActiveFeature from 'components/hooks/useActiveFeature'
 import useCurrentUser from 'components/hooks/useCurrentUser'
 import useNotification from 'components/hooks/useNotification'
 import { downloadCSVFile } from 'components/pages/Bookings/downloadCSVBookings'
-import { DEFAULT_PRE_FILTERS } from 'components/pages/Bookings/PreFilters/_constants'
+import { DEFAULT_PRE_FILTERS } from 'core/Bookings/constants'
 import { TPreFilters } from 'core/Bookings/types'
 import * as pcapi from 'repository/pcapi/pcapi'
 import BookingsScreen from 'screens/Bookings'
